shop/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
ansicolors==1.1.8
anyascii==0.3.2
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
construct==2.5.3
cryptography==44.0.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.17
django-extra-views==0.14.0
django-filter==24.3
django-haystack==3.3.0
django-modelcluster==6.3
django-oscar==3.2.5
django-permissionedforms==0.1
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-taggit==6.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docker==7.1.0
draftjs-exporter==5.0.0
entrypoints==0.4
escapism==1.0.1
et_xmlfile==2.0.0
executing==2.1.0
factory-boy==3.2.1
Faker==33.1.0
fastjsonschema==2.21.1
filetype==1.2.0
frozenlist==1.5.0
gcloud==0.18.3
googleapis-common-protos==1.66.0
halo==0.0.31
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
ipykernel==6.10.0
ipython==8.30.0
ipython-genutils==0.2.0
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==6.1.5
jupyter-highlight-selected-word==0.2.0
jupyter-repo2docker==2024.7.0
jupyter-server==1.15.6
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
l18n==2021.3
laces==0.1.1
log-symbols==0.0.14
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
nbclassic==1.1.0
nbclient==0.7.0
nbconvert==7.16.4
nbformat==5.10.4
nbstripout==0.8.1
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
oauth2client==4.1.3
openpyxl==3.1.5
packaging==24.2
pandocfilters==1.5.1
papermill==2.6.0
parso==0.8.4
pathlib==1.0.1
pefile==2024.8.26
phonenumbers==8.13.51
pillow==11.0.0
pillow_heif==0.21.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
purl==1.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-jwt==4.1.0
python-ptrace==0.9.9
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
semver==3.0.2
Send2Trash==1.8.3
sentry-sdk==2.19.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sorl-thumbnail==12.11.0
soupsieve==2.6
spinners==0.0.24
sqlparse==0.5.2
stack-data==0.6.3
telepath==0.3.1
tenacity==9.0.0
termcolor==2.5.0
terminado==0.18.1
timeago==1.0.16
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wagtail==6.3.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
Willow==1.9.0
yarl==1.18.3
